+++ apps/api/src/golfgenius/golfgenius.module.ts
@@ -11,11 +11,11 @@ import golfGeniusConfig from "./config/golf-genius.config"
 import { validationSchema } from "./config/validation.schema"
 import { GolfgeniusController } from "./golfgenius.controller"
 import { EventSyncService } from "./services/event-sync.service"
+import { ImportAllResultsService } from "./services/import-all-results.service"
 import { IntegrationLogService } from "./services/integration-log.service"
 import { MemberSyncService } from "./services/member-sync.service"
 import { PointsImportService } from "./services/points-import.service"
 import { ProgressTracker } from "./services/progress-tracker"
-import { ResultsImportService } from "./services/results-import.service"
 import { RosterExportService } from "./services/roster-export.service"
 import { RosterPlayerTransformer } from "./services/roster-player-transformer"
 import { ScoresImportService } from "./services/scores-import.service"
@@ -40,10 +40,10 @@ import { ScoresImportService } from "./services/scores-import.service"
 		PointsImportService,
 		ProgressTracker,
 		MemberSyncService,
-		ResultsImportService,
 		RosterExportService,
 		RosterPlayerTransformer,
 		ScoresImportService,
+		ImportAllResultsService,
 	],
 })
 export class GolfgeniusModule {}